Add files via upload
'Library of Peers' is an idea me and 4 colleagues from university came up with for our team project module. It is a website where you can connect with people from your area and borrow or lend books between each other. It uses an existing book ISBN code API which allows you to upload the books which you want to borrow. My main participation in this project was the frontend part and it was both challenging and fun. With no prior knowledge of JS , HTML and CSS , I managed to get my way around learning them through the React framework, thus having contributed to this project with the sign up page , edit profile and add book features.